+++ run.py
@@ -104,13 +104,19 @@ def get_command(machine_name):
                 tasks.heal_code.delay(code)
                 continue
         elif command == 'heal_instances':
-            instance_query = 'where={"type":"express","f5only":false}&max_results=2000'
+            instance_query = 'where={"type":"express"}&max_results=2000'
             instances = utilities.get_eve('sites', instance_query)
             for instance in instances['_items']:
                 tasks.heal_instance.delay(instance)
                 continue
+        elif command == 'heal_instances_no_db':
+            instance_query = 'where={"type":"express"}&max_results=2000'
+            instances = utilities.get_eve('sites', instance_query)
+            for instance in instances['_items']:
+                tasks.heal_instance.delay(instance, db=False, ops=True)
+                continue
         elif command == 'correct_nfs_file_permissions':
-            instance_query = 'where={"type":"express","f5only":false}&max_results=2000'
+            instance_query = 'where={"type":"express"}&max_results=2000'
             instances = utilities.get_eve('sites', instance_query)
             for instance in instances['_items']:
                 tasks.correct_nfs_file_permissions.delay(instance)